Restructure visualiser app
Move files to own visualiser folder
Split into several files, as one app file was getting very large
app.py contains main app that calls everything else
samples, batches, db_view contains the 3 tabs
funcs contains useful functions that are not necessarily linked to tabs
